Optimise script

+++ provision-jenkins-ubuntu-agent.sh
@@ -6,38 +6,35 @@ echo '-----BEGIN RSA PRIVATE KEY-----' | tee /opt/jenkinsssh_id_rsa
 echo $JENKINS_SSH_KEY | sed -e 's/[[:blank:]]\\+/\\n/g' | tee -a /opt/jenkinsssh_id_rsa
 echo '-----END RSA PRIVATE KEY-----' | tee -a /opt/jenkinsssh_id_rsa
 
-apt update -y
+apt-get remove docker docker-engine docker.io containerd runc -y && apt autoremove -y
 
+apt-get update -y && apt-get install -y \
+  ca-certificates \
+  curl \
+  gnupg \
+  gnupg2 \
+  lsb-release
 
 ## debugging why python3-pip sometime can't be found
 cat /etc/apt/sources.list
 apt-cache policy python3-pip
 ## end debug
 
-apt install -y python3-pip
-apt install -y python3-testresources
-apt install -y python2
-pip3 install --upgrade setuptools
-pip3 install --upgrade pip
-pip3 install --upgrade docker-compose
-pip3 install --upgrade virtualenv
-
+curl --silent https://www.postgresql.org/media/keys/ACCC4CF8.asc | apt-key add -
+echo "deb https://apt.postgresql.org/pub/repos/apt/ $(lsb_release -cs)-pgdg main" | tee /etc/apt/sources.list.d/postgresql-pgdg.list > /dev/null
 curl -fsSL https://deb.nodesource.com/setup_14.x | bash -
-apt install -y nodejs
 
-npm install npm@latest minimatch@latest graceful-fs@latest -g
-npm install --global gulp eslint
 
-npm install --global yarn
-
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --batch --yes --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
 echo \
   "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu \
   $(lsb_release -cs) stable" | tee /etc/apt/sources.list.d/docker.list > /dev/null
 
-apt remove docker docker-engine docker.io containerd runc -y && apt autoremove -y
-apt-get update
 curl -sL https://aka.ms/InstallAzureCLIDeb | bash
+
+curl https://packages.microsoft.com/config/ubuntu/21.04/packages-microsoft-prod.deb -o packages-microsoft-prod.deb
+apt-get install ./packages-microsoft-prod.deb
+rm packages-microsoft-prod.deb
 
 # Playwright dependencies. Generated with: npx playwright install-deps
 apt-get install -y --no-install-recommends gstreamer1.0-libav gstreamer1.0-plugins-bad gstreamer1.0-plugins-base \
@@ -51,15 +48,15 @@ apt-get install -y --no-install-recommends gstreamer1.0-libav gstreamer1.0-plugi
   libfontconfig xfonts-cyrillic xfonts-scalable fonts-liberation fonts-ipafont-gothic fonts-wqy-zenhei \
   fonts-tlwg-loma-otf ttf-ubuntu-font-family
 
-apt install -y \
-  ca-certificates \
-  curl \
-  gnupg \
+apt-get update && apt-get install -y \
+  python3-pip \
+  python3-testresources \
+  python2 \
+  nodejs \
   lsb-release \
   openjdk-11-jdk \
   openjdk-17-jdk \
   git \
-  gnupg2 \
   azure-cli \
   docker-ce \
   docker-ce-cli \
@@ -76,9 +73,25 @@ apt install -y \
   libatk-bridge2.0-0 \
   libgtk-3-0 \
   libxss1 \
+  rsync \
+  postgresql \
+  postgresql-contrib \
+  apt-transport-https \
+  dotnet-sdk-5.0 \
+  apt-transport-https \
+  aspnetcore-runtime-5.0 \
+  unzip \
   wget
 
-update-alternatives  --set java /usr/lib/jvm/java-11-openjdk-amd64/bin/java
+pip3 install --upgrade setuptools pip docker-compose virtualenv
+
+npm install npm@latest minimatch@latest graceful-fs@latest -g
+npm install --global \
+  gulp \
+  eslint \
+  yarn
+
+update-alternatives --set java /usr/lib/jvm/java-11-openjdk-amd64/bin/java
 
 #### RVM
 
@@ -93,38 +106,30 @@ rvm install 2.7.6
 
 ####
 
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-apt install -y ./google-chrome-stable_current_amd64.deb
+curl https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -o google-chrome-stable_current_amd64.deb
+apt-get install -y ./google-chrome-stable_current_amd64.deb
+rm -f google-chrome-stable_current_amd64.deb
 
 TFCMT_VERSION=v3.2.1
 curl -fL -o tfcmt.tar.gz https://github.com/suzuki-shunsuke/tfcmt/releases/download/$TFCMT_VERSION/tfcmt_linux_amd64.tar.gz
 tar -C /usr/bin -xzf ./tfcmt.tar.gz tfcmt
 
-wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | apt-key add -
-echo "deb http://apt.postgresql.org/pub/repos/apt/ $(lsb_release -cs)-pgdg main" | tee /etc/apt/sources.list.d/postgresql-pgdg.list > /dev/null
-apt update
-apt install -y postgresql postgresql-contrib
-
-wget https://packages.microsoft.com/config/ubuntu/21.04/packages-microsoft-prod.deb -O packages-microsoft-prod.deb
-dpkg -i packages-microsoft-prod.deb
-rm packages-microsoft-prod.deb
-apt update
-apt install -y apt-transport-https dotnet-sdk-5.0 apt-transport-https aspnetcore-runtime-5.0
-
 [ -e /opt/google/chrome/libosmesa.so ] && rm /opt/google/chrome/libosmesa.so
-LIBOSMESA=$(find / -name 'libOSMesa*' -type f)
+LIBOSMESA=$(find /usr -name 'libOSMesa*' -type f)
 ln -s $LIBOSMESA /opt/google/chrome/libosmesa.so
 echo 'user.max_user_namespaces=10000' > /etc/sysctl.d/90-userspace.conf
 # grubby --args=namespace.unpriv_enable=1 --update-kernel=$(grubby --default-kernel)
-
-apt install -y ruby-full ruby-dev chrony
 
 mkdir /etc/docker && chown -R root:root /etc/docker && chmod 0755 /etc/docker
 echo -e '{\n  \live-restore\: true,\n  \group\: \docker\\n}' > /etc/docker/daemon.conf && chown root:root /etc/docker/daemon.conf && chmod 0644 /etc/docker/daemon.conf
 systemctl enable docker
 
-cp /etc/chrony/chrony.conf{,.orig}
-echo \refclock PHC /dev/ptp0 poll 3 dpoll -2 offset 0\ > /etc/chrony/chrony.conf && cat /etc/chrony/chrony.conf
+# this doesn't exist in a container, flagged so this works when we test the script
+FILE=/etc/chrony/chrony.conf
+if test -f "$FILE"; then
+  cp /etc/chrony/chrony.conf{,.orig}
+  echo \refclock PHC /dev/ptp0 poll 3 dpoll -2 offset 0\ > /etc/chrony/chrony.conf && cat /etc/chrony/chrony.conf
+fi
 
 #Download AzCopy
 wget https://aka.ms/downloadazcopy-v10-linux
@@ -138,14 +143,13 @@ cp ./azcopy_linux_amd64_*/azcopy /usr/bin/
 mkdir /opt/nvm && chown 1001:1001 /opt/nvm
 curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.34.0/install.sh | NVM_DIR=/opt/nvm bash
 
+rm -rf /opt/tfenv /bin/terraform /bin/tfenv
 git clone -b v2.0.0-alpha3 https://github.com/tfutils/tfenv.git /opt/tfenv
 ln -s /opt/tfenv/bin/* /bin
 
-apt install -y unzip
-
 tfenv install 0.13.5 && chown -R 1001:1001 /opt/tfenv
 
-packages=( az azcopy cloud-init docker docker-compose dotnet eslint gcc git google-chrome gulp java /usr/lib/jvm/java-17-openjdk-amd64/bin/java make node npm psql ruby rsync terraform tfcmt tfenv virtualenv yarn wget )
+packages=( az azcopy docker docker-compose dotnet eslint gcc git google-chrome gulp java /usr/lib/jvm/java-17-openjdk-amd64/bin/java make node npm psql ruby rsync terraform tfcmt tfenv virtualenv yarn wget )
 
 for i in "${packages[@]}"
 
@@ -157,5 +161,3 @@ do
      exit 1
   fi
 done
-
-
